refactor(#3280): extract typed-operand tail dispatch from compileBinaryExpression (WAVE C slice 1)
Lift the type-directed tail dispatch of the ~3.1k-LOC god-function
compileBinaryExpression (binary-ops.ts, original lines 1986-3420) into a new
sibling module binary-ops-typed-dispatch.ts / compileTypedBinaryDispatch — the
region reached once both operands are compiled to concrete Wasm value types
(leftType/rightType on the stack): struct-ref valueOf coercion, strict/loose
equality (ref identity, native-string content, $AnyValue tag-aware, externref
abstract-equality cascade), i32/i64/numeric arithmetic, f64 coercion fallback.

- Pure relocation: the helper takes the 14 free setup-locals it reads by value
  and always returns (terminal dispatch, no fall-through, no write-back).
- compileI32BinaryOp / compileI64BinaryOp / compileBooleanBinaryOp promoted to
  export (no emit change); function-level import cycle only.
- binary-ops.ts 4,582 -> 3,164 LOC; compileBinaryExpression ~3,129 -> ~1,710
  LOC; new module 1,485 LOC (< 1,500 threshold).

Byte-identical: prove-emit-identity check IDENTICAL 39/39 (gc/standalone/wasi).
tsc 0, biome lint clean. Smoke test tests/issue-3280.test.ts (#2093 gate) green.


✓
